requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0 